requirements.txt
mediapipe
numpy
opencv
time
websocket-client